Listeners set up for normalizer in FRB.
Listeners are set up in the FRB changes in the js of the colormaps
widget. These include:
- changes in colormaps.cmap_name
- changes in colormaps.is_log
- changes in colormaps.image_array

The colormaps widget holds the normalization routines. However, something
needs to be resolved in the promises because the first render is called
before the normalize function updates the image_array in the colormaps
module.

The array for colormaps.data does not have listeners yet because it
needs to be linked directly to the array in the frb.